+++ test/pivotal.py
@@ -7,3 +7,9 @@ from discopy.pivotal import *
 def test_trace():
     with raises(AxiomError):
         Box('f', 'x', 'y').trace()
+
+
+def test_Sum():
+    f = Box('f', 'x', 'x')
+    assert Sum([f]) == f
+    assert isinstance(f + f, Sum) and (f + f).terms == (f, f)